fix: additional adjustments on pointers

+++ components/ordhook-cli/src/cli/mod.rs
@@ -345,6 +345,9 @@ struct StartCommand {
     /// Check blocks integrity
     #[clap(long = "check-blocks-integrity")]
     pub block_integrity_check: bool,
+    /// Stream indexing to observers
+    #[clap(long = "stream-indexing")]
+    pub stream_indexing_to_observers: bool,
 }
 
 #[derive(Subcommand, PartialEq, Clone, Debug)]
@@ -725,8 +728,11 @@ async fn handle_command(opts: Opts, ctx: &Context) -> Result<(), String> {
                 }
 
                 let ordhook_config = config.get_ordhook_config();
-
-                info!(ctx.expect_logger(), "Starting service...",);
+                let version = env!("GIT_COMMIT");
+                info!(
+                    ctx.expect_logger(),
+                    "Starting service (git_commit = {})...", version
+                );
 
                 let start_block = match cmd.start_at_block {
                     Some(entry) => entry,
@@ -758,7 +764,12 @@ async fn handle_command(opts: Opts, ctx: &Context) -> Result<(), String> {
 
                 let mut service = Service::new(config, ctx.clone());
                 return service
-                    .run(predicates, None, cmd.block_integrity_check)
+                    .run(
+                        predicates,
+                        None,
+                        cmd.block_integrity_check,
+                        cmd.stream_indexing_to_observers,
+                    )
                     .await;
             }
         },
@@ -912,6 +923,16 @@ async fn handle_command(opts: Opts, ctx: &Context) -> Result<(), String> {
             );
             let inscriptions_db_conn_rw =
                 open_readwrite_ordhook_db_conn(&config.expected_cache_path(), ctx)?;
+
+            println!(
+                "{} blocks will be deleted. Confirm? [Y/n]",
+                cmd.end_block - cmd.start_block + 1
+            );
+            let mut buffer = String::new();
+            std::io::stdin().read_line(&mut buffer).unwrap();
+            if buffer.starts_with('n') {
+                return Err("Deletion aborted".to_string());
+            }
 
             delete_data_in_ordhook_db(
                 cmd.start_block,
